Fix param file fixtures and tests
