+++ tests/fixture_plugins/analysis_starter/params_file_content_fixtures.py
@@ -1,74 +1,4 @@
-from pathlib import Path
-
 import pytest
-
-
-@pytest.fixture
-def nextflow_workflow_params_content() -> dict:
-    """Return a dictionary with some parameters for the Nextflow params file."""
-    return {"workflow_param1": "some_value"}
-
-
-@pytest.fixture
-def expected_raredisease_params_file_content(
-    nextflow_case_path: Path,
-    nextflow_sample_sheet_path: Path,
-    raredisease_gene_panel_path2: Path,
-    raredisease_managed_variants_path: Path,
-    nextflow_case_id: str,
-    nextflow_workflow_params_content: dict,
-) -> dict:
-    case_parameters = {
-        "input": nextflow_sample_sheet_path,
-        "outdir": nextflow_case_path,
-        "target_bed_file": "bed_version_file.bed",
-        "analysis_type": "wgs",
-        "save_mapped_as_cram": True,
-        "vcfanno_extra_resources": raredisease_managed_variants_path.as_posix(),
-        "vep_filters_scout_fmt": raredisease_gene_panel_path2.as_posix(),
-        "sample_id_map": Path(
-            nextflow_case_path, f"{nextflow_case_id}_customer_internal_mapping.csv"
-        ),
-    }
-    return case_parameters | nextflow_workflow_params_content
-
-
-@pytest.fixture
-def expected_nallo_params_file_content():
-    return {
-        "filter_variants_hgnc_ids": "/root/case_id/gene_panels.tsv",
-        "input": Path("samplesheet/path"),
-        "outdir": Path("/root/case_id"),
-        "workflow_param1": "some_value",
-    }
-
-
-@pytest.fixture
-def expected_rnafusion_params_file_content(
-    nextflow_case_path: Path,
-    nextflow_sample_sheet_path: Path,
-    nextflow_workflow_params_content: dict,
-) -> dict:
-    """Return a dictionary with parameters for the RNAFUSION params file."""
-    case_parameters = {
-        "input": nextflow_sample_sheet_path,
-        "outdir": nextflow_case_path,
-    }
-    return case_parameters | nextflow_workflow_params_content
-
-
-@pytest.fixture
-def expected_taxprofiler_params_file_content(
-    nextflow_case_path: Path,
-    nextflow_sample_sheet_path: Path,
-    nextflow_workflow_params_content: dict,
-) -> dict:
-    """Return a dictionary with parameters for the Taxprofiler params file."""
-    case_parameters = {
-        "input": nextflow_sample_sheet_path,
-        "outdir": nextflow_case_path,
-    }
-    return case_parameters | nextflow_workflow_params_content
 
 
 @pytest.fixture
